LayoutTests/platform/mac-wk2/plugins/muted-state-expected.txt

Tests that the mute API implementation works as expected. This test needs to be run through WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS plugin1.isMuted is false
PASS plugin1.cachedIsMuted is false
PASS plugin1.isMuted is true
PASS plugin1.cachedIsMuted is true
PASS plugin2.isMuted is true
PASS plugin2.cachedIsMuted is true
PASS successfullyParsed is true

TEST COMPLETE
